requirements.txt
praw
transformers
torch
python-dotenv
tk
fpdf
